requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pydantic==2.6.4
python-dotenv==1.0.1
redis==5.0.4
transformers==4.41.1      
torch==2.3.0              
scikit-learn==1.5.0      
numpy==1.26.4
sqlalchemy==2.0.30
psycopg2-binary==2.9.9    